fix: Fixing the quadrotor model
