--- vehicles/ocean_currents/__init__.py
@@ -1,2 +1,0 @@
-from .abstract_ocean_currents import AbstractOceanCurrents
-from .gaussian_ocean_currents import GaussianOceanCurrents
